requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
mutagen==1.45.1
pycparser==2.21
pycryptodomex==3.14.1
PyNaCl==1.4.0
six==1.16.0
typing_extensions==4.2.0
websockets==10.3
yarl==1.7.2
yt-dlp==2022.4.8
